requirements.txt
torch
torchvision
matplotlib
torchviz
gradio
scikit-learn
langchain==0.1.8
langchain-core>=0.1.24,<0.2
langchain-google-genai>=0.0.6
langchain-community>=0.0.21,<0.1
python-dotenv==1.0.1
tavily-python>=0.3.0
google-generativeai>=0.3.0